Replaced ClientConfig/Builder interfaces with concrete classes
