--- src/main/java/com/appsflyer/donkey/client/ClientConfigBuilder.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 2020 AppsFlyer
- *
- * Licensed under the Apache License, Version 2.0 (the "License")
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package com.appsflyer.donkey.client;
-
-import io.vertx.core.Vertx;
-import io.vertx.ext.web.client.WebClientOptions;
-
-public interface ClientConfigBuilder {
-  
-  static ClientConfigBuilder create() {
-    return new ClientConfigImpl.ClientConfigBuilderImpl();
-  }
-  
-  ClientConfigBuilder vertx(Vertx vertx);
-  
-  ClientConfigBuilder clientOptions(WebClientOptions clientOptions);
-  
-  ClientConfigBuilder debug(boolean val);
-  
-  ClientConfig build();
-}
